+++ Apermo/Sniffs/PHP/ForbiddenObjectCastSniff.php
@@ -1,4 +1,6 @@
 <?php
+declare(strict_types=1);
+
 /**
  * Discourage (object) casts that produce stdClass instances.
  *